backend/requirements.txt
absl-py==2.1.0
accelerate==1.4.0
adapters==1.1.0
affine @ file:///home/conda/feedstock_root/build_artifacts/affine_1674245120525/work
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1696634205638/work
anthropic==0.42.0
anyio==3.7.1
argcomplete==3.3.0
arxiv==2.0.0
astunparse==1.6.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1704011227531/work
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
black==24.10.0
blinker==1.9.0
boto3==1.35.83
botocore==1.35.83
bulldozer-dtm==1.0.1
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
click-plugins==1.1.1
cligj @ file:///home/conda/feedstock_root/build_artifacts/cligj_1633637764473/work
cmake==3.29.3
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.7
datasets==3.3.2
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==7.2.1
e2b==1.0.5
e2b-code-interpreter==1.0.3
exiv2==0.16.3
fastapi==0.104.1
feedparser==6.0.10
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
fiona==1.10.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
GDAL @ file:///Users/runner/miniforge3/conda-bld/gdal-split_1711284165673/work/build/swig/python
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
gradio==5.12.0
gradio_client==1.5.4
greenlet==3.1.1
grpcio==1.64.1
gspread==6.1.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.0
idna==3.7
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kiwisolver==1.4.8
langchain==0.3.13
langchain-anthropic==0.3.1
langchain-community==0.3.13
langchain-core==0.3.27
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.4
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.4
libclang==18.1.1
litellm==1.57.5
llama-cloud==0.1.12
llama-cloud-services==0.6.0
llama-index==0.12.16
llama-index-agent-openai==0.4.5
llama-index-cli==0.4.0
llama-index-core==0.12.16.post1
llama-index-embeddings-huggingface==0.5.1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-ollama==0.5.2
llama-index-llms-openai==0.3.18
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.0
llama_cpp_python==0.3.7
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.10.0
mdurl==0.1.2
memory-profiler==0.61.0
ml-dtypes==0.4.1
morecantile @ file:///home/conda/feedstock_root/build_artifacts/morecantile_1707500365838/work
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
neo4j-driver==5.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225421156/work/dist/numpy-1.26.4-cp312-cp312-macosx_11_0_arm64.whl#sha256=011d57633d659db8280a5811d62d0c5a615719a6d7cbd46a04f6ffce0c2a7db3
oauthlib==3.2.2
ollama==0.4.7
openai==1.59.6
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.12
packaging==24.1
pandas==2.1.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
peft==0.14.0
pgvector==0.3.6
pillow==11.0.0
pinecone==5.4.2
pinecone-client==5.0.1
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pinecone-plugin-records==1.1.0
platformdirs==4.3.6
primp==0.10.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pymongo==4.6.0
pyogrio==0.10.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1709721012883/work
pypdf==5.3.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyproj @ file:///Users/runner/miniforge3/conda-bld/pyproj_1702028165637/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.1
rasterio @ file:///Users/runner/miniforge3/conda-bld/rasterio_1702440567149/work
rasterstats==0.20.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rio-cogeo @ file:///home/conda/feedstock_root/build_artifacts/rio-cogeo_1712066419199/work
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
ruff==0.9.1
s3transfer==0.10.4
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.5.0
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==69.5.1
sgmllib3k==1.0.0
shapely==2.0.4
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
smolagents==1.2.1
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.6
SQLAlchemy==2.0.23
starlette==0.27.0
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.4.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.4
transformers==4.47.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.24.0
uvloop==0.21.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
